remove Float64. Double is enough

+++ src/main/java/org/psjava/formula/geometry/PointByDistanceAzimuth.java
@@ -2,14 +2,13 @@ package org.psjava.formula.geometry;
 
 import org.psjava.ds.geometry.Azimuth;
 import org.psjava.ds.geometry.Point2D;
-import org.psjava.ds.numbersystrem.Float64;
 
 public class PointByDistanceAzimuth {
 
-	public static Point2D<Float64> calc(double d, Azimuth a) {
+	public static Point2D<Double> calc(double d, Azimuth a) {
 		double x = d * Math.cos(a.radian());
 		double y = d * Math.sin(a.radian());
-		return Point2D.create(Float64.valueOf(x), Float64.valueOf(y));
+		return Point2D.create(x,y);
 	}
 
 	private PointByDistanceAzimuth() {
