+++ src/test/java/org/psjava/formula/geometry/DistanceBetweenPointsTest.java
@@ -3,8 +3,6 @@ package org.psjava.formula.geometry;
 import org.junit.Assert;
 import org.junit.Test;
 import org.psjava.ds.geometry.Point2D;
-import org.psjava.ds.numbersystrem.Float64;
-import org.psjava.formula.geometry.DistanceBetweenPoints;
 
 public class DistanceBetweenPointsTest {
 
@@ -19,8 +17,8 @@ public class DistanceBetweenPointsTest {
 		Assert.assertEquals(5, DistanceBetweenPoints.calc(toPoint(0, 10), toPoint(3, 14)), 1e-10);
 	}
 
-	private Point2D<Float64> toPoint(int x, int y) {
-		return Point2D.create(Float64.valueOf(x), Float64.valueOf(y));
+	private Point2D<Double> toPoint(double x, double y) {
+		return Point2D.create(x,y);
 	}
 
 }
